requirements.txt
timm
pyyaml 
tqdm
scikit-learn
matplotlib
pandas
Pillow
seaborn
torch
torchvision
zipfile